server/requirements.txt
request
djangorestframework